Renamed 'internet' module to 'net'
