+++ etherparse/src/net/ip_number_impl.rs
