+++ etherparse/src/net/ip_payload_slice.rs
