+++ etherparse/src/net/ipv6_header_slice.rs
