+++ etherparse/src/net/ipv6_slice.rs
