+++ etherparse/src/net/lax_ipv6_slice.rs
